+++ src/finance/programgarden_finance/ls/indtp/t1514/blocks.py
@@ -63,7 +63,7 @@ from typing import Dict, List, Literal, Optional
 from pydantic import BaseModel, Field, PrivateAttr
 from requests import Response
 
-from ....models import BlockRequestHeader, BlockResponseHeader, SetupOptions
+from ...models import BlockRequestHeader, BlockResponseHeader, SetupOptions
 
 
 class T1514RequestHeader(BlockRequestHeader):